fix gradient computation
The problem
-----------

radial gradients:
The center of a circle with diameter 4 is at (2,2). This is not the center
of the pixel at the coordinates (2,2) because the pixel coordinates are the
top left corner of the pixel's square. Therefore the center of pixel (2,2)
is at (2.5,2.5). The same goes for all other pixels.

linear gradients:
Analogous to the radial gradients the distance of a pixel to an edge is not
x or (w-1)-x or y or (h-1)-y but x+0.5 or w-(x+0.5) or y+0.5 or h-(y+0.5).

The solution
------------

The distance computation now uses the pixel center
(pixel coordinate + (0.5,0.5)) for computation.
Radial gradients are now centred properly.
Linear gradients don't need to reduce the width or height anymore.