Add to the package heading explaining the nature of the canvas types.
The canvas primitive types are used throughout the codebase, but they're
not flagged as being non-interactive and designed to be extended through
widgets.

This change adds a note to the package overview leaving a breadcrumb for
attentive readers of the documentation to look to Widgets when attempting
to extend functionality.