file dialog: use ./ as start dir by default
This swaps ~/ and ./ as the default directory.

I have also made / the failover if both of those error. That *shouldn't*
happen though.

Note that / should be safe even on Windows, since NT supports
POSIX-style paths. `/` should expand to NTFS's root, which is almost
always `C:\`.

XXX: might be worth setting the failover dir in the platform driver. `/`
is probably not a good failover choice for iOS, although I'm not sure
the file dialog even makes sense there, since they have their own system
one? I'm not an iOS dev, so I'm not sure.