file dialog: Add ShowFile*At methods
This adds the concept of a "starting directory", which may be set with a
new setter method. If there is a problem with this directory, or it is
an empty string, then the CWD is preferred, followed by the user home,
followed by `/`.

This also adds two new utility methods, ShowFileOpenAt, and
ShowFileSaveAt which allow the starting directory to be explicitly set
as arguments.

This should not break the existing API.