Use icons for file dialog favourites (#1186)
This PR adds icons to the file picker favourites buttons to make it look a bit better and to make it possible for the user to know what a favourite points to without reading the full name of it. To make that happen, I had to add icons for download, computer, document and storage.