SplitContainer: improve computeSplitLengths()
This also fixes the math regarding splitting the size.
When the length (width or height) of the container is 100
and the divider has a thickness of 8, then only 92 are available for the content.
When the offset is 0.25 then the leading content should
be a third of the size of the trailing.
With the numbers from above, that would mean 23 and 69. But with the old implementation it was 21 and 71.