add storage/rfc3986
This will be immediately reverted, but if we decide to revive it for
some reason, it might be a good starting point. I started writing this
and then realized Go already has an RFC3986 implementation here:
https://golang.org/pkg/net/url/

I plan to use that unless there is some reason to build out or own
instead.