first-pass implementation of FileRepository
This still does not implement CopyableRepository or MoveableRepository.

We may want to look at the URI parsing, and put some of the smarts used
for path separator handling into a `CanonicalRepository` for it.

We also still need to decide where/how the driver will register it to
the `file://` repository during start-up.

I also have no idea if this even works or compiles, haven't tested it
yet at all.