Update platform and mobile tests to run Go 1.16
This makes sure that we run tests on the latest stable Go release.
This has the added benefit of making builds slightly waster was well, thanks to the improved linker.
"Linking is 20-25% faster than 1.15 and requires 5-15% less memory on average for linux/amd64, with larger improvements for other architectures and OSes."