fix backspace on android for AOSP-based keyboards
Issue #1941

Gboard always sends backspace events, however, other keyboards only send
or enable the backspace based on the state of the textentry. Also, some
don't send the event, but instead modify the textentry text itself.

This edits the entry to always show some text, a zero. This way the
keyboards kill be able to edit/backspace characters. The textChanged
then sends a backspace Key Event to the Go code.

In the case of using a keyboard (such as Gboard) that sends its own
backspace event, we need a way to stop processing of two deletes (the
one Gboard sends, and the one textChanged sends). A backspaceDelay timer
is used to only process backspaces a certain time (50ms) apart.