Handle resize of windows asynchronously.
This is what glfw API actually expect. In most case, you are going to
get the window size you asked for and it will "work" to resize the
canvas ahead of the window being resized by the compositor, but in the
odd case it isn't or on platform where it could be denied, asynchronous
behavior is a must.